// src/exercises/dto/exercise.fill.db.dto.ts
export class ExerciseFillDBDTO{
      problem: string;
      dificult: number;
      type: number;
      number_variables: number;

    /*  fraction: boolean;
      root: boolean;
      ln: boolean;
      log: boolean;
      tang: boolean;
      sen: boolean;
      cosn: boolean;
      arctang: boolean;
      arcsen: boolean;
      arccosn: boolean;
      cotang: boolean;
      exp: boolean;*/

      frac : boolean;
    //  plus : boolean;
    //  minus : boolean;
      cdot : boolean;

      sqrt : boolean;
      sqrtn : boolean;
      arcos : boolean;
      arsin :boolean;
      arctan :boolean;
      cos :boolean;
      csc :boolean;
      cosh :boolean;
      ln :boolean;
      log :boolean;
      sin :boolean;
      tan :boolean;
      sec :boolean;
      exp :boolean;
      ele : boolean;

      int_def : boolean;
   //   createdAt: Date;
    
}